Clean up coverage for admin-tool-ext-win extension (#10806)
