+++ extensions/admin-tool-ext-win/coverConfig.json
@@ -4,7 +4,8 @@
 	"relativeCoverageDir": "../../coverage",
 	"ignorePatterns": [
 		"**/node_modules/**",
-		"**/test/**"
+		"**/test/**",
+		"main.js"
 	],
 	"reports": [
 		"cobertura",
